+++ .eslintrc.js
@@ -1,4 +1,5 @@
 const rules = {
+  'no-lonely-if': 'off',
   'no-bitwise': 'off',
   'import/no-extraneous-dependencies': 'off',
   'no-nested-ternary': 'off',
@@ -131,6 +132,7 @@ module.exports = {
         'prefer-destructuring': ['off'],
         'no-restricted-globals': ['off'],
         'vue/multi-word-component-names': 'off',
+        'vue/no-deprecated-slot-attribute': 'off',
       },
     },
     {
@@ -142,6 +144,7 @@ module.exports = {
         'import/extensions': ['off'],
         'no-restricted-globals': ['off'],
         'vue/multi-word-component-names': 'off',
+        'vue/no-deprecated-slot-attribute': 'off',
       },
     },
     // SVELTE KITCHEN_SINK